made new curation pipeline into one
